Refactor into BaseServerHandler and BaseEchoServerHandler classes
