+++ examples/ssl_echo_server.py
@@ -49,6 +49,8 @@ def main() -> None:
         pool.setup()
         while True:
             time.sleep(1)
+    except KeyboardInterrupt:
+        pass
     finally:
         pool.shutdown()
